front: move test components to own dir, fix sonarlint errors
