+++ frontend/src/features/functions/components/index.ts
@@ -1,13 +1,13 @@
 import ExternalSalesDataImport from "./ExternalSalesDataImport";
 import ThirtyYearErrorModal from "./ThirtyYearErrorModal";
-import ThirtyYearErrorSelect from "./ThirtyYearErrorSelect";
-import ThirtyYearErrorTest from "./ThirtyYearErrorTest";
-import ThirtyYearErrorTestResult from "./ThirtyYearErrorTestResult";
 import ThirtyYearLoadedResults from "./ThirtyYearLoadedResults";
 import ThirtyYearResultListItem from "./ThirtyYearResultListItem";
 import ThirtyYearResults from "./ThirtyYearResults";
 import ThirtyYearSkippedList from "./ThirtyYearSkippedList";
 import ThirtyYearSkippedListItem from "./ThirtyYearSkippedListItem";
+import ThirtyYearErrorSelect from "./ThirtyYearTests/ThirtyYearErrorSelect";
+import ThirtyYearErrorTest from "./ThirtyYearTests/ThirtyYearErrorTest";
+import ThirtyYearErrorTestResult from "./ThirtyYearTests/ThirtyYearErrorTestResult";
 
 export {
     ExternalSalesDataImport,
